Refine SelectoComponents execution seams
