+++ lib/selecto_components/execution/executor.ex
@@ -4,8 +4,8 @@ defmodule SelectoComponents.Execution.Executor do
   """
 
   alias SelectoComponents.Execution.Result
+  alias SelectoComponents.Execution.QueryHelpers
   alias SelectoComponents.Form
-  alias SelectoComponents.Form.ParamsState
   alias SelectoComponents.Performance.MetricsCollector
   alias SelectoComponents.Views.Aggregate.Options, as: AggregateOptions
   alias SelectoComponents.Views.Detail.Options, as: DetailOptions
@@ -13,7 +13,12 @@ defmodule SelectoComponents.Execution.Executor do
   @spec run(SelectoComponents.Execution.Plan.t(), Phoenix.LiveView.Socket.t()) :: Result.t()
   def run(plan, socket) do
     {query_result, view_meta, page_query_cache} =
-      ParamsState.execute_query_for_plan(plan.selecto, plan.params, plan.view_meta, socket)
+      QueryHelpers.execute_query_with_pagination(
+        plan.selecto,
+        plan.params,
+        plan.view_meta,
+        socket
+      )
 
     case query_result do
       {:ok, {rows, columns, aliases}, metadata} ->
@@ -64,10 +69,10 @@ defmodule SelectoComponents.Execution.Executor do
     end
 
     {rows_for_display, view_meta} =
-      ParamsState.cap_aggregate_rows_for_result(rows, view_meta, plan.params)
+      QueryHelpers.maybe_cap_aggregate_rows(rows, view_meta, plan.params)
 
     normalized_rows =
-      ParamsState.normalize_rows_for_result(
+      QueryHelpers.normalize_rows_for_view(
         rows_for_display,
         columns,
         socket.assigns.view_config.view_mode
@@ -82,7 +87,7 @@ defmodule SelectoComponents.Execution.Executor do
       if AggregateOptions.aggregate_view_mode?(plan.params), do: page_query_cache, else: nil
 
     cache_debug_info =
-      ParamsState.build_query_cache_debug_info_for_result(
+      QueryHelpers.build_query_cache_debug_info(
         detail_cache_assignment,
         plan.params,
         normalized_rows,
@@ -118,8 +123,8 @@ defmodule SelectoComponents.Execution.Executor do
       field_filters: Selecto.filters(plan.selecto),
       presentation_context: plan.presentation_context,
       query_results: {normalized_rows, columns, aliases},
-      used_params: ParamsState.drop_runtime_only_params_public(plan.params),
-      applied_view: ParamsState.get_map_value_public(plan.params, :view_mode),
+      used_params: drop_runtime_only_params(plan.params),
+      applied_view: get_map_value(plan.params, :view_mode),
       view_meta: view_meta,
       detail_page_cache: detail_cache_assignment,
       aggregate_page_cache: aggregate_cache_assignment,
@@ -133,7 +138,7 @@ defmodule SelectoComponents.Execution.Executor do
     sanitized_error =
       Form.sanitize_error_for_environment(
         error,
-        ParamsState.execution_error_opts_public(error, plan.params, operation: "view-apply")
+        execution_error_opts(error, plan.params, operation: "view-apply")
       )
 
     {error_sql, error_params} = safe_to_sql(plan.selecto)
@@ -144,8 +149,8 @@ defmodule SelectoComponents.Execution.Executor do
       field_filters: Selecto.filters(plan.selecto),
       presentation_context: plan.presentation_context,
       query_results: nil,
-      used_params: ParamsState.drop_runtime_only_params_public(plan.params),
-      applied_view: ParamsState.get_map_value_public(plan.params, :view_mode),
+      used_params: drop_runtime_only_params(plan.params),
+      applied_view: get_map_value(plan.params, :view_mode),
       view_meta: view_meta,
       detail_page_cache:
         if(DetailOptions.detail_view_mode?(plan.params), do: page_query_cache, else: nil),
@@ -161,7 +166,7 @@ defmodule SelectoComponents.Execution.Executor do
     sanitized_error =
       Form.build_selecto_error(:query_error, inspect(error), %{original_error: error})
       |> Form.sanitize_error_for_environment(
-        ParamsState.execution_error_opts_public(error, plan.params, operation: "view-apply")
+        execution_error_opts(error, plan.params, operation: "view-apply")
       )
 
     {error_sql, error_params} = safe_to_sql(plan.selecto)
@@ -172,8 +177,8 @@ defmodule SelectoComponents.Execution.Executor do
       field_filters: Selecto.filters(plan.selecto),
       presentation_context: plan.presentation_context,
       query_results: nil,
-      used_params: ParamsState.drop_runtime_only_params_public(plan.params),
-      applied_view: ParamsState.get_map_value_public(plan.params, :view_mode),
+      used_params: drop_runtime_only_params(plan.params),
+      applied_view: get_map_value(plan.params, :view_mode),
       view_meta: view_meta,
       detail_page_cache:
         if(DetailOptions.detail_view_mode?(plan.params), do: page_query_cache, else: nil),
@@ -204,4 +209,84 @@ defmodule SelectoComponents.Execution.Executor do
   end
 
   defp list_field(_map, _key), do: []
+
+  defp execution_error_opts(error, params, extra_opts) do
+    view_mode = get_map_value(params, :view_mode)
+
+    Keyword.merge(
+      [
+        stage: execution_error_stage(error),
+        category: execution_error_category(error),
+        code: execution_error_code(error),
+        view_mode: view_mode
+      ],
+      extra_opts
+    )
+  end
+
+  defp execution_error_stage(%{__struct__: module} = error) when module == Selecto.Error do
+    case Map.get(error, :type, :query_error) do
+      :validation_error -> :input
+      :configuration_error -> :configuration
+      :field_resolution_error -> :configuration
+      :timeout_error -> :timeout
+      :connection_error -> :db_execute
+      :transformation_error -> :result_process
+      :query_error -> if(Map.get(error, :query), do: :db_execute, else: :query_build)
+      _ -> :unknown
+    end
+  end
+
+  defp execution_error_stage(:timeout), do: :timeout
+  defp execution_error_stage({:error, :timeout}), do: :timeout
+  defp execution_error_stage(_), do: :db_execute
+
+  defp execution_error_category(%{__struct__: module} = error) when module == Selecto.Error do
+    case Map.get(error, :type, :query_error) do
+      :validation_error -> :validation
+      :configuration_error -> :configuration
+      :field_resolution_error -> :configuration
+      :timeout_error -> :timeout
+      :connection_error -> :connection
+      :transformation_error -> :processing
+      :permission_error -> :authorization
+      :query_error -> :query
+      _ -> :runtime
+    end
+  end
+
+  defp execution_error_category(:timeout), do: :timeout
+  defp execution_error_category({:error, :timeout}), do: :timeout
+  defp execution_error_category(_), do: :query
+
+  defp execution_error_code(%{__struct__: module} = error) when module == Selecto.Error do
+    case Map.get(error, :type, :query_error) do
+      :validation_error -> :validation_error
+      :configuration_error -> :invalid_view_config
+      :field_resolution_error -> :unknown_field
+      :timeout_error -> :query_timed_out
+      :connection_error -> :connection_error
+      :transformation_error -> :result_processing_failed
+      :permission_error -> :permission_error
+      :query_error -> if(Map.get(error, :query), do: :db_query_failed, else: :query_build_failed)
+      type -> type
+    end
+  end
+
+  defp execution_error_code(:timeout), do: :query_timed_out
+  defp execution_error_code({:error, :timeout}), do: :query_timed_out
+  defp execution_error_code(_), do: :db_query_failed
+
+  defp drop_runtime_only_params(params) when is_map(params),
+    do: Map.delete(params, "_presentation_context")
+
+  defp drop_runtime_only_params(params), do: params
+
+  defp get_map_value(map, key, default \\ nil)
+
+  defp get_map_value(map, key, default) when is_map(map) do
+    Map.get(map, key, Map.get(map, to_string(key), default))
+  end
+
+  defp get_map_value(_map, _key, default), do: default
 end
